index1.php
Hey Tara, I'm not entirely sure how to use GitHub and I wasn't able to edit the codes directly but I was able to copy the code and edit it from another source. This is what I have so far. I'm sorry if it's not much. I have been having trouble figuring out how to use HTML but I think I'm getting the hang of it. I will continue to work on the web page and try to get a little more done now that I have a better understanding of how to use HTML. Like I said before I wasn't entirely sure how to use this website so I am sending the edited code to you through the proposed file changes since this is really the only thing that I can access. Let me know what you think and if you have any suggestions.